kern: make struct kmem_cache fit into two cache lines
Previously, the size of struct kmem_cache was 136 bytes, just eight
bytes larger than 128 bytes, which is typically two cache lines on
today's CPUs.

By reducing the size of the name field which holds a human-readable
description by eight bytes to 24 bytes, the struct kmem_cache can be
made fit into two cache lines.  This change should not affect the
usefulness of this field.  For reference, the length of the largest
hard-coded name is 17.

* kern/slab.h (KMEM_CACHE_NAME_SIZE): Define to 24.