kern: fix formatting of multiboot modules
Previously, bootstrap_create would print the multiboot modules with
padding applied to the end of the line.  As multiboot modules as used
by the Hurd span more than one line.  This makes the list of modules
hard to read and it looks unclean, more like an accident.
Furthermore, it is not clear what the intend of this was, as the
padding is applied at the end of the line, with no further information
printed thereafter.

* kern/bootstrap.c (bootstrap_create): Remove variable maxlen and len,
update printfs.