include: do not guard the host_slab_info RPC with MACH_VM_DEBUG
Previously, the definition of the host_slab_info RPC was guarded with
MACH_VM_DEBUG, even though it is not at all concerned with the VM
subsystem.  Furthermore, there was no "skip" directive for
host_slab_info.

The function host_slab_info is guarded with MACH_DEBUG.  The server
for the RPCs in mach_debug.defs is only used if MACH_DEBUG is
defined. There is no need to guard host_slab_info.

* include/mach_debug/mach_debug.defs (host_slab_info): Unconditionally
include the RPC.