include: fix the embedded type definitions in memory_object.defs
In order to use MIG translation functions to lookup memory objects,
preprocessor macros have been introduced into the definition of
memory_object_t in 50cc5152.

The procedure definitions contain inlined type definitions in order to
change the type of the argument in question (i.e. to make it
polymorphic).  The inline definitions however lack the destructor
function, leading to reference leaks when a reference is acquired in
the intran function.

* include/mach/memory_object.defs: Add the destructor functions to the
inlined type definitions.