doc: restore section `Inherited Ports'
Previously, the section `Inherited Ports' was commented out.  This was
done, as the functionality was unused by the Hurd.  The functions
`mach_ports_register' and `mach_ports_lookup' were never removed, and
are exposed to user space.

This patch brings the documentation back and adds a remark at the top,
that the section documents the original intentions for this interface.

I chose bringing back the documentation over removing the
functionality because I like to make use of it as a method for service
discovery that is deliberately orthogonal to the way the service
lookup is usually done in the Hurd.  This can be used to implement
robust low-level debugging facilities.

* doc/mach.texi: Restore section `Inherited Ports'.