device: fix net_rcv_msg-messages
Previously, all net_rcv_msg-messages sent by net_deliver were
malformed.  It never was a problem in practice, since the messages are
not complex and thus the kernel does not try to parse the message.

struct net_rcv_msg contains an additional field of type boolean_t.
This field has no associated type descriptor, so it must not be
included in the message.

* device/net_io.c (net_deliver): Account for the extra field in the
msgh_size calculation.