ipc: tune size of cached kernel message buffers
The previous limit was 256 bytes.  That seems a little crummy by
todays standards, and we are frequently sending bigger packets
(e.g. every RPC containing a string_t on Hurd).

Use the page size for IKM_SAVED_KMSG_SIZE to make sure the page is
pinned to a single processor.

* ipc/ipc_kmsg.h (IKM_SAVED_KMSG_SIZE): Define to `PAGE_SIZE'.