kern: avoid #if 0ing out thread_collect_scan
Currently, `thread_collect_scan' does nothing because `pcb_collect' is
a nop.  Its body is exempt from compilation by means of the
preprocessor.

This is unfortunate as it increases the risk of bitrot, and we still
need to pay the price of rate-limiting thread_collect_scan.

* kern/thread.c (thread_collect_scan): Drop #if 0 around the body.
* vm/vm_pageout.c (vm_pageout_scan): Do not call
`consider_thread_collect' and document why.