Fix interactivity of inactive threads
A new thread will mosty probably start working, assume it will take
its share of CPU, to avoid having to find it out slowly.  Decaying will
however fix that quickly if it actually does not work.

This fixes stalling issues when a program keeps creating threads.

* kern/thread.c (thread_create): Set new_thread's cpu_usage and
sched_usage to a fair share of the current load.