smp: Add pseudoclass to manage APIC operations
The SMP support requires access, register and configure some APIC structures,
like Local APIC and IOAPIC.
This pseudoclass includes functions to register some APIC structures into the kernel,
and access to these structures to extract some information.

*apic.h: Header file recovered from Mach 4 source code, and updated for xAPIC.
Includes some structs with Local APIC and IOAPIC fields,
and the declaration of the functions.

*apic.c: Source file. Includes the definition of the functions, within some globals
of apic_info structures and the lapic reference.