i386 pmap: Omit pmap workaround on i486 or later.
As described in XXX comments, the workaround for memory mapping is
implemented for 80386 and it is unnecessary on i486 or later.  Thus, it
is safe to omit that if the kernel is built for the recent (1989~)
processors.

Fuhito Inagawa pointed out the problem to me.

* i386/i386/trap.c (kernel_trap): Disable the workaround when the kernel
is built for i[456]86.
* i386/intel/pmap.c (pmap_protect, pmap_enter): Ditto.
* i386/intel/read_fault.c: Define intel_read_fault if and only if it is
necessary.