hack vm memory object proxy creation for vm arrays
* vm/memory_object_proxy.c: truncate vm array types as if they were
  the rpc_ version because MIG can't handle that. This rpc can't
  handle more than one element anyway.

Note that the same issue with vm arrays is present at least with
syscall emulation, but that functionality seems unused for now.

A better fix could be to add a vm descriptor type in include/mach/message.h,
but then probably we don't need to use the rpc_ types in MIG anymore,
they would be needed only for the syscall definitions.

Signed-off-by: Luca Dariz <luca@orpolo.org>
Message-Id: <20220628101054.446126-15-luca@orpolo.org>