gnumach: Remove PCI_CFG2_* as a reserved ioport range
PCI CONF2 access method was deprecated in PCI v2.0

Also, the conf2 ioport range is only activated for use
as pci configuration space when conf1 registers are accessed
in a particular sequence.

Since the pci-arbiter does not access the conf1 forwarding register,
we are safe to remove this completely from gnumach to allow
normal I/O use of the conf2 range as modern hardware expects.
Message-Id: <20220823032638.292813-1-damien@zamaudio.com>