interrupt: Fix saving irq/ipl when linux drivers are disabled
When Linux drivers are disabled, in hardclock() the linux_timer_intr()
call is dropped, and gcc can tail-recursion-optimize the call to
clock_interrupt(). To do so, it overwrites the hardclock() parameters to
suit the clock_interrupt parameters layout. This however means it thrashes
the backups that the interrupt() function had made of irq/ipl, leading to
mayhem.

interrupt should thus really properly separate its irq/ipl backups from the
interrupt function parameters.

Thanks a lot to Etienne Brateau for the tricky investigation!