Update mach_debug interfaces to use struct.
The new interfaces will be compatible both with a 64 bits kernel and
userland and 64 bits kernel and 32 bit userland. Also removed many
of the uses of natural_t where an unsigned int suffices. Ideally we
should replace natural_t with something more portable such as uintptr_t
or a basic int type for counters since for the most part we don't
need counters to have the same width as the pointer type.

* i386/include/mach/i386/vm_types.h: Define rpc_unsigned_long.
* include/mach/mach_types.defs: Define type rpc_vm_offset_t.
* include/mach_debug/hash_info.h: Use unsigned int instead of natural_t.
* include/mach_debug/mach_debug_types.defs: Update cache_info_t,
  hash_info_bucket_t, vm_region_info_t, vm_object_info_t and
  vm_page_info_t to use struct.
* include/mach_debug/slab_info.h: Use rpc_vm_size_t and
  rpc_unsigned_long for compatibility with 32 bits userland.
* include/mach_debug/vm_info.h: Update struct fields to use the correct
  types and avoid natural_t whenever we just want to keep a count of
  something.
Message-Id: <Y32lp1SuV01ImCx9@viriathus>