vm_wire_all: Fix vm_map_protect case
If a "wire_required" process calls vm_map_protect(0), the
memory gets unwired as expected. But if the process then calls
vm_map_protect(VM_PROT_READ) again, we need to wire that memory.

(This happens to be exactly what glibc does for its heap)

This fixes Hurd hangs on lack of memory, during which mach was swapping
pieces of mach-defpager out.