Remove bootstrap.defs
As far as I can see, this file was imported in the very beginning of GNU Mach
history, and unused since then. Nobody implements or uses this interface. GNU
Mach uses a different way to pass the privileged ports to the bootstrap tasks:
instead of the task(s) actively asking for the ports in an RPC, the ports are
preemptively inserted into the IPC space(s) of the task(s), as configured by
the boot script.

Remove bootstrap.defs so as to not confuse anyone into thinking it works.
Message-Id: <20230319151017.531737-3-bugaevc@gmail.com>