x86_64: Fix updating fsgs base on context switch
Commit 31dd30a94a682955c3c9e2f42252b4a07687067a "add setting gs/fsbase"
added the code to set fs and gs bases on context_switch. However, this
was only being done when switching context via the explicit
switch_context() method, but not in other cases where the context is
switched, such as with call_continuation().

Instead, put setting fsgs base into switch_ktss(), where it will be
called in all cases.
Message-Id: <20230511135844.837338-1-bugaevc@gmail.com>