Fix interrupt handling
Logic for interrupts:

- interrupt.S raises spl (thus IF cleared)
- interrupt.S EOI
- interrupt.S calls the handler
   - for pure in-kernel handlers, they do whatever they want with IF
     cleared.
   - when a userland handler is registers, queue_intr masks the irq.
- interrupt.S lowers spl with splx_cli, thus IF still cleared
- iret, that sets IF

- later on, userland acks the IRQ, that unmasks the irq

The key to this change is that all interrupts, including IPIs, are
treated the same way. Eg. the spl level is now raised before an IPI and
restored after. Also, EOI is not needed inside irq_acknowledge.

With this change and the experimental change not to dispatch threads
direct to idle processors in the scheduler, I no longer observe kernel
faults, but an occasional hang does occur.

Message-Id: <20231002033906.124427-1-damien@zamaudio.com>