smp: Remove hardcoded AP_BOOT_ADDR
This took some time to figure out.
Involves a hand-crafted 16 bit assembly instruction [1]
because it requires an immediate for the memory address
of far jump.  This required self-modifying code
to inject the next instruction, therefore I added a near
jump to clear the instruction cache queue in case the pipeline
cached the unmodified jump location.

[1] Intel Architecture Software Developer's Manual,
    Volume 2: Instruction Set Reference Manual